Core: remove obsolete window observers (#76)
* Core: remove obsolete window observers

Comming probably from Firefox Boostrap extension days, the windows
observers where not used correctly - as sometimes they were looking of
chrome.windows object and sometimes on the extension background window
(depending on the extension access to chrome.windows).

Additionaly, those were the subject of the race condition in case
browser was starting in the backround.

The only current usecase was to check if the browser runs in incognito
mode, but even that was not correct as all modules should do the check
based on chrome.tabs, chrome.webNavigation or chrome.webRequest and
prevent any processing if incognito mode was detected. Which is very
much the case.

This all makes this code safe to be removed.